# MapStr accessor optimisations