@ph 2018-04-12
LGTM, the code is simplier than before 👍  waiting for Jenkins run to complete. The Travis faillure seems to be related a docker container on Metricbeat and not on this code.